pytorch-flask-deploy/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
Pillow==7.2.0
requests==2.24.0
torch==1.6.0+cpu
torchvision==0.7.0+cpu
urllib3==1.25.10
Werkzeug==1.0.1
